KVM: x86: Change vCPU's register cache avail/dirty variable to unsigned long
Change 'regs_avail' and 'regs_dirty' member of 'struct kvm_vcpu_arch'
from u32 to unsigned long as they are passed as bitmap to test_bit(),
which is an atomic operation, which requires the address passed to it to
be aligned to size of unsigned long.

With CONFIG_DEBUG_ATOMIC=y and CONFIG_DEBUG_ATOMIC_LARGEST_ALIGN=y,
running a simple guest on one Intel platform triggered couple of WARNING
splats in instrument_atomic_check_alignment().  E.g., one of them was:

  ------------[ cut here ]------------
  (unsigned long)v & mask
  WARNING: ./include/linux/instrumented.h:67 at vmx_load_mmu_pgd+0x2d9/0x2f0 [kvm_intel], CPU#170: CPU 0/KVM/5875
  ...
  RIP: 0010:vmx_load_mmu_pgd+0x2d9/0x2f0 [kvm_intel]
  ...
  Call Trace:
   <TASK>
   kvm_mmu_load+0x360/0xc50 [kvm]
   ...
   vcpu_enter_guest+0xf25/0x1500 [kvm]

They were caused by the test_bit() when kvm_register_is_available() and
kvm_register_is_dirty() pass the per-vCPU 'regs_avail' and 'regs_dirty'
respectively as bitmap to test_bit().  Since both 'regs_avail' and
'regs_dirty' are u32, they are not aligned to size of unsigned long,
causing WARNING splats like the above one in test_bit().

The alignment requirement of the atomic operation is to make sure on x86
an atomic instruction can never access data crossing cacheline boundary,
which can cause severe performance degradation.  If the kernel has split
lock detection enabled, this kind access can cause #AC split lock
exception and crashing the kernel.

Fix by changing both of them to unsigned long.  This wastes 4-bytes for
each vCPU but is pretty straightforward.  It's also feasible to combine
'regs_avail' and 'regs_dirty' to a single unsigned long, but this will
require logic change to the code, i.e., a 32-bit offset needs to be
added to the testing bit when using one of them.

Note:

- No real split lock exception was triggered but only the WARNING splats
  (the kernel enabled the split lock detection to "warn" mode).

- The 'regs_avail' and 'regs_dirty' were introduced long before split
  lock detection was merged, hence no 'Fixes' tag.

+++ arch/x86/include/asm/kvm_host.h
@@ -799,8 +799,8 @@ struct kvm_vcpu_arch {
 	 * kvm_{register,rip}_{read,write} functions.
 	 */
 	unsigned long regs[NR_VCPU_REGS];
-	u32 regs_avail;
-	u32 regs_dirty;
+	unsigned long regs_avail;
+	unsigned long regs_dirty;
 
 	unsigned long cr0;
 	unsigned long cr0_guest_owned_bits;
